+++ app/src/main/java/com/telen/easylineup/data/Player.kt
@@ -11,8 +11,8 @@ import androidx.room.*
         ]
 )
 data class Player(
-        @PrimaryKey(autoGenerate = true) var id: Int,
-        @ColumnInfo(name = "teamID") var teamId: Int,
+        @PrimaryKey(autoGenerate = true) var id: Long,
+        @ColumnInfo(name = "teamID") var teamId: Long,
         @ColumnInfo(name = "name") var name: String,
         @ColumnInfo(name = "shirtNumber") var shirtNumber: Int,
         @ColumnInfo(name = "licenseNumber") var licenseNumber: Long)
