+++ app/src/main/java/com/telen/easylineup/data/PlayerDao.kt
@@ -17,11 +17,11 @@ interface PlayerDao {
     fun deletePlayer(player: Player)
 
     @Query("SELECT * from players WHERE id = :playerID" )
-    fun getPlayerById(playerID: Int): LiveData<Player>
+    fun getPlayerById(playerID: Long): LiveData<Player>
 
     @Query("SELECT * FROM players")
     fun getPlayers(): LiveData<List<Player>>
 
     @Query("SELECT players.* FROM players INNER JOIN teams ON players.teamID = teams.id WHERE teams.id = :teamId")
-    fun getPlayersForTeam(teamId: Int): LiveData<List<Player>>
+    fun getPlayersForTeam(teamId: Long): LiveData<List<Player>>
 }
